Minor variable type Changes, removed unnecessary file reading and texture loading

+++ Chess v5.0/src/Other/BoardStats.cpp
@@ -302,16 +302,17 @@ void BoardStats::DisplayNewFENMessage(const string& fen) {
 }
 std::string BoardStats::format(int number) {
     std::string str = std::to_string(number);
-    int len = static_cast<int>(str.length());
+    size_t len = str.length();
     if (len <= 3) // No need for formatting if the number is less than 1000
         return str;
 
-    int commas = (len - 1) / 3; // Calculate the number of commas to be inserted
+    size_t commas = (len - 1) / 3; // Calculate the number of commas to be inserted
     std::string formatted;
     formatted.reserve(len + commas);
 
-    int i = 0;
-    int j = len % 3 == 0 ? 3 : len % 3; // Adjust for the first segment
+    size_t i = 0;
+    size_t j = len % 3 == 0 ? 3 : len % 3; // Adjust for the first segment
+
     formatted.append(str, 0, j);
 
     for (; i + j < len; i += j, j = 3) {
